refactor: BIOINFO-31 rename exclude-mnps step to split-mnps
Renamed the subworkflow exclude_mnps to handle_mnps to better reflect its functionality.
This step, separates MNPs into individual SNPs. Only invalid MNPs are removed

Updated the documentation, which incorrectly stated that this step was required for the VQSR step.
It is, in fact, required for the entire joint genotyping procedure.
The step can be skipped for efficiency only if it is certain that there are no MNPs in the input gVCF.

Change the parameter name controlling the execution for this step from exclude_mnps to skip_handle_mnps.

Other: correct wrong column samplesheet column name in the documentation for the phenopacket file